+++ crates/mimic/src/data/executor/load/dynamic.rs
@@ -2,7 +2,7 @@ use crate::{
     Error,
     data::{
         DataError,
-        executor::{DebugContext, Loader, with_resolver},
+        executor::{DebugContext, Loader},
         query::{LoadFormat, LoadQueryDyn},
         response::{LoadCollectionDyn, LoadResponse},
         store::{DataStoreRegistry, IndexStoreRegistry},
@@ -71,7 +71,7 @@ impl LoadExecutorDyn {
         self.debug.println(&format!("query.load_dyn: {query:?}"));
 
         // resolver
-        let resolved_entity = with_resolver(|r| r.entity(E::PATH))?;
+        // let resolved_entity = resolve_entity::<E>()?;
 
         // do we include a row?
         fn include_row(row: &DataRow, query: &LoadQueryDyn, path: &str) -> bool {
@@ -87,7 +87,7 @@ impl LoadExecutorDyn {
         // but we have to filter by the fn above and paginate
         let loader = Loader::new(self.data_reg, self.index_reg, self.debug);
         let rows = loader
-            .load(&resolved_entity, &query.selector, None)?
+            .load::<E>(&query.selector, None)?
             .into_iter()
             .filter(|row| include_row(row, &query, E::PATH))
             .skip(query.offset as usize)