+++ crates/mimic/src/types/mod.rs
@@ -1,5 +1,7 @@
+mod def;
 mod error;
 mod prim;
 
+pub use def::*;
 pub use error::ErrorTree;
 pub use prim::*;