+++ crates/mimic_design/src/imp/entity.rs
@@ -1,6 +1,7 @@
 use crate::{
     imp::{Imp, Implementor},
     node::{Entity, MacroNode, Trait},
+    traits::Schemable,
 };
 use mimic::schema::types::Cardinality;
 use proc_macro2::TokenStream;
@@ -15,14 +16,16 @@ pub struct EntityKindTrait {}
 impl Imp<Entity> for EntityKindTrait {
     fn tokens(node: &Entity, t: Trait) -> Option<TokenStream> {
         // store
-        let store = node.store;
+        let store = &node.store;
         let mut q = quote! {
-            const STORE: &'static str = #store;
+            const STORE: &'static str = stringify!(#store);
         };
 
-        q.extend(searchable_fields(node));
-        q.extend(sort_key(node));
+        //     q.extend(indexes(node));
+        q.extend(key(node));
+        q.extend(values(node));
         q.extend(build_sort_key(node));
+        //    q.extend(build_index_key(node));
 
         let tokens = Implementor::new(&node.def, t)
             .set_tokens(q)
@@ -32,22 +35,55 @@ impl Imp<Entity> for EntityKindTrait {
     }
 }
 
-// searchable_fields
-fn searchable_fields(node: &Entity) -> TokenStream {
-    let entries = node.fields.iter().filter_map(|field| {
+// indexes
+fn indexes(node: &Entity) -> TokenStream {
+    let defs = node.indexes.iter().map(|ei| ei.schema());
+
+    quote! {
+        const INDEXES: &'static [::mimic::schema::node::EntityIndex] = &[
+            #(#defs),*
+        ];
+    }
+}
+
+// key
+fn key(node: &Entity) -> TokenStream {
+    let fields = node.sort_keys.iter().map(|sk| {
+        let field_ident = &sk.field;
+
+        quote!(self.#field_ident.to_string())
+    });
+
+    quote! {
+        fn key(&self) -> ::mimic::types::Key {
+            Key(vec![
+                #(#fields),*
+            ])
+        }
+    }
+}
+
+// values
+fn values(node: &Entity) -> TokenStream {
+    let inserts = node.fields.iter().filter_map(|field| {
         let field_ident = &field.name;
         let field_name = field.name.to_string();
-        let item = &field.value.item;
 
         match field.value.cardinality() {
             Cardinality::One => Some(quote! {
-                (#field_name.to_string(), <#item as ::mimic::traits::FieldQueryable>::to_query_value(&self.#field_ident))
+                if let Some(v) = self.#field_ident.to_query_value() {
+                    map.insert(#field_name.into(), Some(v));
+                }
             }),
 
             Cardinality::Opt => Some(quote! {
-                (#field_name.to_string(), self.#field_ident
-                    .as_ref()
-                    .and_then(<#item as ::mimic::traits::FieldQueryable>::to_query_value))
+                if let Some(inner) = &self.#field_ident {
+                    if let Some(v) = inner.to_query_value() {
+                        map.insert(#field_name.into(), Some(v));
+                    }
+                } else {
+                    map.insert(#field_name.into(), None);
+                }
             }),
 
             Cardinality::Many => None,
@@ -55,74 +91,66 @@ fn searchable_fields(node: &Entity) -> TokenStream {
     });
 
     quote! {
-        fn searchable_fields(&self) -> ::std::collections::HashMap<String, Option<String>> {
-            [
-                #(#entries),*
-            ].into_iter().collect()
+        fn values(&self) -> ::std::collections::HashMap<String, Option<String>> {
+            use ::mimic::traits::FieldQueryable;
+
+            let mut map = ::std::collections::HashMap::with_capacity(3);
+            #(#inserts)*
+
+            map
         }
     }
 }
 
-// sort_key
-fn sort_key(node: &Entity) -> TokenStream {
-    let fields = node.sort_keys.iter().map(|sk| {
-        let field_ident = &sk.field;
+// build_sort_key
+fn build_sort_key(node: &Entity) -> TokenStream {
+    let mut index: usize = 0;
 
-        quote!(&self.#field_ident.to_string())
-    });
+    // set_fields
+    let set_fields = node.sort_keys.iter().filter_map(|sort_key| {
+        let field = sort_key.field.as_ref()?;
+        let i = index;
+        index += 1;
 
-    quote! {
-        fn sort_key(&self) -> Result<::mimic::data::store::SortKey, Error> {
-            Self::build_sort_key(&[
-                #(#fields),*
-            ])
-        }
-    }
-}
-
-// format_sort_key
-fn format_sort_key(node: &Entity) -> TokenStream {
-    let fields = node
-        .sort_keys
-        .iter()
-        .filter_map(|sk| sk.field.clone())
-        .collect::<Vec<_>>();
-
-    // Prepare the quote for setting struct fields based on the provided values slice
-    let set_fields = fields.iter().enumerate().map(|(i, ident)| {
-        quote! {
+        Some(quote! {
             if let Some(value) = values.get(#i) {
-                this.#ident = value.parse().unwrap_or_default();
+                this.#field = value.parse().unwrap_or_default();
             }
-        }
+        })
     });
 
-    // quote for generating the output vector using the ORM trait to
-    // format each field as a sort key
-    let format_keys = fields.iter().map(|ident| {
-        quote! {
-            ::mimic::traits::FieldSortKey::format(&this.#ident)
-        }
+    // format_keys
+    let format_keys = node.sort_keys.iter().filter_map(|sort_key| {
+        let field = sort_key.field.as_ref()?;
+        let path_str = field.to_string();
+
+        Some(quote! {
+            ::mimic::data::types::SortKeyPart::new(
+                #path_str,
+                this.#field.to_sort_key_part(),
+            )
+        })
     });
 
-    // create inner
-    let inner = if fields.is_empty() {
+    let inner = if node.sort_keys.is_empty() {
         quote!(Vec::new())
     } else {
         quote! {
-            let mut this = Self::default();
+            use ::mimic::traits::FieldSortKey;
 
+            let mut this = Self::default();
             #(#set_fields)*
 
-            // Collect formatted keys and then take only as many as there are input values
             let format_keys = vec![#(#format_keys),*];
             format_keys.into_iter().take(values.len()).collect()
         }
     };
 
     quote! {
-        fn build_sort_key(values: &[::std::string::String]) -> ::mimic::data::store::SortKey {
-            #inner
+        fn build_sort_key(values: &[::std::string::String]) -> ::mimic::data::types::SortKey {
+            ::mimic::data::types::SortKey::from_parts({
+                #inner
+            })
         }
     }
 }
@@ -147,17 +175,17 @@ impl Imp<Entity> for EntitySearchTrait {
               match field.value.cardinality() {
                     Cardinality::One => quote! {
                         ( #name_str, |s: &#ident, text|
-                            ::mimic::traits::FormatQueryValue::contains_text(&s.#name, text)
+                            ::mimic::traits::FieldQueryable::contains_text(&s.#name, text)
                         )
                     },
                     Cardinality::Opt => quote! {
                         ( #name_str, |s: &#ident, text|
-                            s.#name.as_ref().map_or(false, |v| ::mimic::traits::FormatQueryValue::contains_text(v, text))
+                            s.#name.as_ref().map_or(false, |v| ::mimic::traits::FieldQueryable::contains_text(v, text))
                         )
                     },
                     Cardinality::Many => quote! {
                         ( #name_str, |s: &#ident, text|
-                             s.#name.iter().any(|v| ::mimic::traits::FormatQueryValue::contains_text(v, text))
+                             s.#name.iter().any(|v| ::mimic::traits::FieldQueryable::contains_text(v, text))
                         )
                     },
                 }
@@ -212,24 +240,24 @@ impl Imp<Entity> for EntitySortTrait {
 
             asc_fns.extend(quote! {
                 fn #asc_fn(a: &#node_ident, b: &#node_ident) -> ::std::cmp::Ordering {
-                    ::mimic::traits::Orderable::cmp(&a.#field_ident, &b.#field_ident)
+                    ::mimic::traits::FieldOrderable::cmp(&a.#field_ident, &b.#field_ident)
                 }
             });
 
             desc_fns.extend(quote! {
                 fn #desc_fn(a: &#node_ident, b: &#node_ident) -> ::std::cmp::Ordering {
-                    ::mimic::traits::Orderable::cmp(&b.#field_ident, &a.#field_ident)
+                    ::mimic::traits::FieldOrderable::cmp(&b.#field_ident, &a.#field_ident)
                 }
             });
 
             match_arms.extend(quote! {
-                (#field_str, ::mimic::data::SortDirection::Asc) => comps.push(#asc_fn),
-                (#field_str, ::mimic::data::SortDirection::Desc) => comps.push(#desc_fn),
+                (#field_str, ::mimic::data::types::SortDirection::Asc) => comps.push(#asc_fn),
+                (#field_str, ::mimic::data::types::SortDirection::Desc) => comps.push(#desc_fn),
             });
         }
 
         let q = quote! {
-            fn sort(order: &[(String, ::mimic::data::SortDirection)])
+            fn sort(order: &[(String, ::mimic::data::types::SortDirection)])
                 -> Box<dyn Fn(&#node_ident, &#node_ident) -> ::std::cmp::Ordering>
             {
                 #asc_fns