+++ tests/src/lib.rs
@@ -1,7 +1,0 @@
-#![feature(test)]
-
-pub mod bench;
-
-extern crate test;
-
-use test::Bencher;